requirements.txt
openai>=1.3.0
pandas>=2.2.0
tqdm>=4.66.0
pyyaml>=6.0.1
python-docx>=1.1.0
pypdf>=4.2.0
pdfminer.six>=20231228
rapidfuzz>=3.4.0
rank_bm25>=0.2.2
jinja2>=3.1.3
nltk>=3.8.1
